+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the values for the breakpoints and the associated docs.
 - Updates the `viewport` preview options in Storybook to align with the Reservoir breakpoints.
 
+## Prerelease
+
+### Adds
+
+- Adds the `"editorMode"` option (pencil) to the `Icon` component.
+
 ## 3.0.0 (March 14, 2024) React 18 / Chakra 2.8
 
 ### Adds